app/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-environ==0.4.5
django-filter==2.0.0
django-storages==1.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
